Removes wrappers around upstream Dqlite app and options. We don't need to add this complexity.
The build directives have been removed. We will just always assume that
we are building with CGo.
